+++ main.php
@@ -47,6 +47,8 @@ $showIcon = tpl_getConf('showIcon');
 
 		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
 		<script src="https://cdn.jsdelivr.net/npm/flatpickr"></script>
+
+		<?php $translation = plugin_load('helper','translation'); ?>
 	</head>
 
 	<body class="docs ">
@@ -64,9 +66,15 @@ $showIcon = tpl_getConf('showIcon');
 						// get logo either out of the template images folder or data/media folder
 						$logoSize = array();
 						$logo = tpl_getMediaFile(array(':wiki:logo.png', ':logo.png', 'images/logo.png', 'images/logo.svg', ':wiki:dokuwiki-128.png'), false, $logoSize);
+						if ($translation) {
+							$link = $translation->buildTransID($conf['lang'], $conf['start']);
+							$mainpage_link = ltrim($link[0], ':');
+						} else {
+							$mainpage_link = '';
+						}
 						// display logo and wiki title in a link to the home page
 						tpl_link(
-							wl(),
+							wl($mainpage_link),
 							'<img class="logo" src="'.$logo.'" alt="" /><span>'.$conf['title'].'</span>',
 							'accesskey="h" title="[H]"'
 						);
@@ -141,7 +149,7 @@ $showIcon = tpl_getConf('showIcon');
 							<div class="mx-auto" style="max-width: fit-content;">
 							<?php
 							if (!empty($_SERVER['REMOTE_USER'])) {
-								echo '<a href="'.wl('', 'do=profile').'" id="user-info" class="nav-item nav-link"> ';
+								echo '<a href="'.wl($mainpage_link, 'do=profile').'" id="user-info" class="nav-item nav-link"> ';
 								tpl_userinfo();
 								echo '</a>';
 							}
@@ -178,7 +186,15 @@ $showIcon = tpl_getConf('showIcon');
 							<div id="dokuwiki__aside" class="ct-toc-item active">
 								<div class="leftsidebar">
 									<div class="ct-toc-link">
-										<?php tpl_link(wl('sidebar', '', true), "Содержание", 'class="wikilink1"'); ?>
+										<?php
+											if ($translation) {
+												$link = $translation->buildTransID($conf['lang'], $conf['sidebar']);
+												$sidebar_link = ltrim($link[0], ':');
+											} else {
+												$sidebar_link = $conf['sidebar'];
+											}
+											tpl_link(wl($sidebar_link, '', true), $lang['toc'], 'class="wikilink1"');
+										?>
 									</div>
 									<?php tpl_includeFile('sidebarheader.html')?>
 									<?php tpl_include_page($conf['sidebar'], 1, 1)?>
@@ -224,6 +240,10 @@ $showIcon = tpl_getConf('showIcon');
 									?>
 								</ul>
 							</div> -->
+
+							<?php
+							if ($translation) echo $translation->showTranslations();
+							?>
 
 						</nav>
 					</div>